refactor: 유효성 검증 함수 메뉴별로 분리
