+++ src/assets/validations/manageTapValidation.js
@@ -75,20 +75,6 @@ const checkAddProductsQuantityInput = quantity => {
   return isError;
 };
 
-const isSoldOut = quantity => {
-  if (quantity === 0) {
-    alert(ERROR_MESSAGE.soldOut);
-    return true;
-  }
-};
-
-const isLackMoney = (money, price) => {
-  if (money < price) {
-    alert(ERROR_MESSAGE.lackMoney);
-    return true;
-  }
-};
-
 export const checkAddProductsInputs = (machine, name, price, quantity) => {
   let isValid = true;
   if (checkAddProductsNameInput(machine, name)) {
@@ -96,43 +82,6 @@ export const checkAddProductsInputs = (machine, name, price, quantity) => {
   } else if (checkAddProductsPriceInput(price)) {
     isValid = false;
   } else if (checkAddProductsQuantityInput(quantity)) {
-    isValid = false;
-  }
-
-  return isValid;
-};
-
-export const checkChargeChangeInput = input => {
-  let isValid = true;
-  if (isEmptyInput(input)) {
-    isValid = false;
-  } else if (isNegative(input)) {
-    isValid = false;
-  } else if (isNotDivideByTen(input)) {
-    isValid = false;
-  }
-
-  return isValid;
-};
-
-export const checkInsertMoneyInput = input => {
-  let isValid = true;
-  if (isEmptyInput(input)) {
-    isValid = false;
-  } else if (isNegative(input)) {
-    isValid = false;
-  } else if (isNotDivideByTen(input)) {
-    isValid = false;
-  }
-
-  return isValid;
-};
-
-export const checkCanPurchase = (quantity, money, price) => {
-  let isValid = true;
-  if (isSoldOut(parseInt(quantity))) {
-    isValid = false;
-  } else if (isLackMoney(money, parseInt(price))) {
     isValid = false;
   }
